// AngularSPA/ClientApp/app/components/event/bookevent/event.d.ts
//export interface Welcome {
//    statusCode: number;
//    isSuccess: boolean;
//    description: null;
//    data: Data;
//}

export class EventData {
    eventId: number;
    createTime: string;
    strCreatedTime: string;
    invitations: Invitation[];
};

export class Invitation {
    id: number;
    name: string;
    email: string;
    mobile: string;
    content: string;
    status: number;
    createTime: string;
    strCreatedTime: null;
    requestSentTime: null;
    strRequestSentTime: null;
    responseTime: null;
    strResponseTime: null;
};